revert: custom entropy pool
Seemed like a good idea but further tests revealed that it only behaved well with low buffer sizes. As the size of the given buffer increased, the performance of Scramble degraded way faster than the equivalent rand.Read call did.
