+++ core/crypto_test.go
@@ -73,6 +73,22 @@ func TestCompare(t *testing.T) {
 	}
 }
 
+func TestScramble(t *testing.T) {
+	b := make([]byte, 32)
+	Scramble(b)
+	if bytes.Equal(b, make([]byte, 32)) {
+		t.Error("buffer not scrambled")
+	}
+	c := make([]byte, 32)
+	Scramble(c)
+	if bytes.Equal(b, make([]byte, 32)) {
+		t.Error("buffer not scrambled")
+	}
+	if bytes.Equal(b, c) {
+		t.Error("random repeated")
+	}
+}
+
 func TestHash(t *testing.T) {
 	known := make(map[string]string)
 	known[""] = "DldRwCblQ7Loqy6wYJnaodHl30d3j3eH+qtFzfEv46g="
